+++ src/components/auth/forms/sign-up-image-selector.tsx
