+++ src/components/auth/forms/sign-up-steps/index.ts
