+++ src/components/auth/forms/sign-up-steps/step-1-nickname.tsx
@@ -3,7 +3,7 @@ import { Hand, Check, Loader2 } from 'lucide-react';
 import { useState, useRef, useEffect } from 'react';
 import { cn } from '@/components/common/ui/(shadcn)/lib/utils';
 import { BaseInput } from '@/components/common/ui/input';
-import SignupImageSelector from '@/components/forms/sign-up-image-selector';
+import SignupImageSelector from '@/components/auth/forms/sign-up-image-selector';
 import { useDebounce } from '@/hooks/common/use-debounce';
 import { useNicknameCheckQuery } from '@/hooks/queries/auth/use-nickname-check';
 import { StepHeader } from './step-header';
