+++ src/components/auth/forms/sign-up-steps/step-4-study-format-types.tsx
