+++ src/components/auth/forms/sign-up-steps/step-6-success.tsx
