+++ src/components/auth/forms/sign-up-steps/step-header.tsx
